requirements.txt
numpy==1.23.4
opencv_python==4.6.0.66
pandas==1.5.2
pymongo==4.3.3
PyYAML==6.0
